add CommandPanel component and CommandDictionary stub
pull dictionary logic out from SelectedCommand into SequenceEditor for reuse
